Split Public Key from Token retrieval
